Game engine extended
Actor convention is intorduced. (A similar one like in unreal engine)
ActorCollection is introduced.
Displaying object logic is redesigned according to actor and actor collection classes.

Everything displayed on the screen is considered as an actor.
Every actor is drawable and has two methods to help game design which are atStart and tick.
atStart method is called at the beginning of the game for every actor.
tick method called at every frame of the game for every actor.
This frame based approach is very similar to unreal engine and it helps to create the game easily.